chore: split up decoder aux column builders into separate files (#1233)
